Add edit group settings form
